refactor(genja-core): Split settings schema into section modules.
